+++ src/pss_fleet_data/__init__.py
@@ -1,17 +1,19 @@
 from pssapi.entities import Alliance as PssAlliance
 from pssapi.entities import User as PssUser
 
-from . import core, models
+from . import core, models, utils
 from .client import PssFleetDataClient
 from .core import exceptions
-from .models import Collection, CollectionMetadata
+from .models import Collection, CollectionMetadata, enums
 
 
 __all__ = [
     # Modules
     core.__name__,
+    enums.__name__,
     exceptions.__name__,
     models.__name__,
+    utils.__name__,
     # Classes
     Collection.__name__,
     CollectionMetadata.__name__,
