+++ src/pss_fleet_data/core/__init__.py
@@ -1,9 +1,9 @@
-from . import config, enums, exceptions, utils
+from .. import utils
+from . import config, exceptions
 
 
 __all__ = [
     config.__name__,
-    enums.__name__,
     exceptions.__name__,
     utils.__name__,
 ]
