+++ tests/10_core/utils/test_remove_timezone.py
@@ -3,7 +3,7 @@ from datetime import datetime
 import pytest
 from dateutil.parser import parse as parse_datetime
 
-from pss_fleet_data.core.utils import remove_timezone
+from pss_fleet_data.utils import remove_timezone
 
 
 test_cases_invalid = [
