+++ tests/30_client/10_helper/helper_test_cases.py
@@ -3,7 +3,7 @@ from datetime import datetime, timezone
 
 import pytest
 
-from pss_fleet_data.core.enums import ParameterInterval
+from pss_fleet_data.models.enums import ParameterInterval
 
 
 _DEFAULT_API_KEY = os.getenv("TEST_DEFAULT_API_KEY")
